move resolveStylesAndClasses to a separate file
